+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
 * Remove wrapper `openssl::_EVP_PKEY_CTX_get_rsa_oaep_label`. This is
   technically an ABI break, but since the wrappers are not considered part of
   the public API, we do not bump the SOVERSION for this.
+* Improve error message in MoCOCrWException that is thrown in case of invalid
+  signature validation.
 
 ## Fixed